perf: 修改创建authbook信号
