+++ apps/assets/signals_handler/__init__.py
@@ -1,4 +1,5 @@
-from .common import *
+from .asset import *
+from .system_user import *
 from .authbook import *
 from .node_assets_amount import *
 from .node_assets_mapping import *
