+++ apps/assets/signals_handler/authbook.py
@@ -3,9 +3,12 @@ from django.apps import apps
 from simple_history.signals import pre_create_historical_record
 from django.db.models.signals import post_save, pre_save
 
-from ..models import AuthBook
+from common.utils import get_logger
+from orgs.utils import tmp_to_root_org
+from ..models import AuthBook, SystemUser
 
 AuthBookHistory = apps.get_model('assets', 'HistoricalAuthBook')
+logger = get_logger(__name__)
 
 
 @receiver(pre_create_historical_record, sender=AuthBookHistory)
@@ -22,6 +25,25 @@ def pre_create_historical_record_callback(sender, instance=None, history_instanc
             setattr(history_instance, attr, system_user_attr_value)
 
 
+@receiver(post_save, sender=AuthBook)
+def on_authbook_post_create(sender, instance, **kwargs):
+    # 去掉这个资产的管理用户
+    if instance.systemuser and instance.systemuser.is_admin_user:
+        with tmp_to_root_org():
+            deleted_count, other = AuthBook.objects.filter(
+                asset=instance.asset,
+                systemuser__type=SystemUser.Type.admin
+            ).exclude(id=instance.id).delete()
+            logger.debug('Remove asset old admin user: {}'.format(deleted_count))
+
+
 @receiver(pre_save, sender=AuthBook)
-def on_authbook_create_update_version(sender, instance, **kwargs):
+def on_authbook_pre_create(sender, instance, **kwargs):
+    # 升级版本号
     instance.version = instance.history.all().count() + 1
+
+
+@receiver(post_save, sender=AuthBook)
+def on_authbook_create(sender, instance=None, created=True, **kwargs):
+    if not created:
+        return
